+++ sparsemv.c
@@ -14,14 +14,14 @@
  * @param y Return vector
  * @return int 0 if no error
  */
-int sparsemv(struct mesh *A, const double * const x, double * const y)
+int sparsemv(struct mesh *A, const float * const x, float * const y)
 {
   const int nrow = (const int) A->local_nrow;
 
   #pragma omp parallel for
   for (int i=0; i<nrow; i++) {
-      double sum = 0.0;
-      const double * const cur_vals = (const double * const) A->ptr_to_vals_in_row[i];
+      float sum = 0.0;
+      const float * const cur_vals = (const float * const) A->ptr_to_vals_in_row[i];
       const int * const cur_inds = (const int * const) A->ptr_to_inds_in_row[i];
       const int cur_nnz = (const int) A->nnz_in_row[i];
       int j;
